+++ cargo-workspaces/src/utils/mod.rs
@@ -1,3 +1,4 @@
+mod basic_checks;
 mod cargo;
 mod changable;
 mod config;
@@ -7,8 +8,10 @@ mod error;
 mod git;
 mod list;
 mod pkg;
+mod publish;
 mod version;
 
+pub use basic_checks::basic_checks;
 pub use cargo::{cargo, cargo_config_get, change_versions, rename_packages};
 pub use changable::{ChangeData, ChangeOpt};
 pub use config::{read_config, PackageConfig, WorkspaceConfig};
@@ -17,8 +20,11 @@ pub use dev_dep_remover::{should_remove_dev_deps, DevDependencyRemover};
 pub(crate) use error::{debug, info, warn};
 pub use error::{get_debug, set_debug, Error};
 pub use git::{git, GitOpt};
-pub use list::{list, ListOpt};
+pub use list::{list, ListOpt, ListPublicOpt};
 pub use pkg::{get_pkgs, is_private, Pkg};
+pub use publish::{
+    create_http_client, filter_private, is_published, package_registry, RegistryOpt,
+};
 pub use version::VersionOpt;
 
 pub type Result<T = ()> = std::result::Result<T, Error>;